Structure of code changed
- Collision with other objects is dealt with in different scripts
