+++ Assets/Scripts/eatSweet.cs
