+++ Assets/Scripts/objectCollisionRock.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 4cc4769feabfc45618bd2602073a6cd7
-timeCreated: 1493198759
+guid: 12046150cc8e14d6e9d25bdbfb1c295c
+timeCreated: 1493559415
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
