--- Assets/createGridObjects.cs
@@ -1,16 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class createGridObjects : MonoBehaviour {
-	[HideInInspector]public List<Vector2> gridPoints;
-	// Use this for initialization
-	void Start () {
-		
-	}
-	
-	// Update is called once per frame
-	void Update () {
-		
-	}
-}
