--- Assets/spawnGridPoints.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: f2bc4594197c94fb38c468cd282bb409
-timeCreated: 1493469375
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
